+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.0] - 2018-08-02
+
+### Added
+
+- New trait `HasResponseFactory` used by many middlewares that need to configure the PSR-17 response factory.
+- New trait `HasStreamFactory` used by many middlewares that need to configure the PSR-17 stream factory.
+
 ## [2.0.0] - 2018-08-01
 
 ### Added
@@ -193,6 +200,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Stream factory
 
 
+[2.1.0]: https://github.com/middlewares/utils/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/middlewares/utils/compare/v1.2.0...v2.0.0
 [1.2.0]: https://github.com/middlewares/utils/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/middlewares/utils/compare/v1.0.0...v1.1.0